backend/requirements.txt
# Core
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Async & Queue
celery==5.3.6
redis==5.0.7
kombu==5.3.4
flower==2.0.1

# Email Validation
dnspython==2.6.1
aiosmtplib==3.0.1
email-validator==2.1.1
pyIsEmail==1.0.6

# HTTP & Async
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.3

# Stripe
stripe==10.6.0

# Storage & CSV
boto3==1.34.129
pandas==2.2.2
python-magic==0.4.27

# Security & Auth
cryptography==42.0.8
PyJWT==2.8.0
django-ratelimit==4.1.0

# API Docs
drf-spectacular==0.27.2

# Environment
python-dotenv==1.0.1
python-decouple==3.8

# Monitoring
sentry-sdk==2.6.0
django-health-check==3.18.3

# Utilities
Pillow==10.3.0
celery-redbeat==2.2.0
django-celery-results==2.5.1
django-celery-beat==2.6.0
whitenoise==6.7.0
gunicorn==22.0.0
